requirements.txt
fastapi
uvicorn
google-generativeai
beautifulsoup4
requests
pydantic
